feat: add zh_CN as locale, refactor sort by views, re-organize library modules
