+++ src/library/sorting/sort-by-index/index.js
@@ -1,0 +1,28 @@
+export class SortByIndexStrategy {
+  /**
+   * Sorts a list of videos by their index
+   * @param {Array<Element>} videos
+   * @param {"asc" | "desc"} sortOrder
+   * @returns {Array<Element>}
+   */
+  sort(videos, sortOrder) {
+    return Array.from(videos)
+      .slice(0, 100)
+      .sort((videoA, videoB) => {
+        const indexA = videoA.querySelector(
+          "yt-formatted-string#index"
+        ).innerText;
+        const indexB = videoB.querySelector(
+          "yt-formatted-string#index"
+        ).innerText;
+
+        if (sortOrder === "asc") {
+          return Number(indexA) - Number(indexB);
+        }
+
+        if (sortOrder === "desc") {
+          return Number(indexB) - Number(indexA);
+        }
+      });
+  }
+}
