+++ src/library/sorting/sort-by-views/parsers/base.js
@@ -1,0 +1,6 @@
+export class ViewsParser {
+  /** @param {Element} videoInfo */
+  parse(videoInfo) {
+    throw new Error("Parser not implemented");
+  }
+}
